requirements.txt
apistar==0.5.40
aiofiles==0.3.2
requests==2.9.1
apache-libcloud==2.3.0